+++ jscomp/test/dist-es6/jscomp/test/es6_tests/es6_import.mjs
@@ -3,8 +3,4 @@
 import * as Es6_export from "./es6_export.mjs";
 
 console.log(Es6_export.default);
-
-export {
-  
-}
 /*  Not a pure module */