+++ test/blackbox-tests/as-ppx-config-attribute.t
@@ -14,5 +14,7 @@
   
   const x = 1;
   
-  exports.x = x;
+  module.exports = {
+    x,
+  }
   /* No side effect */