+++ test/blackbox-tests/set-package-name-in-js.t
@@ -51,5 +51,7 @@ B depends on A, so it should import a.js in the right path
   
   const t = A.t;
   
-  exports.t = t;
+  module.exports = {
+    t,
+  }
   /* No side effect */